feat: Add feature to write directory listing to file
In this commit, a capability has been added to write the output of the directory listing to a file. This feature enables the user to save the file and directory structure for future reference. The changes include creating a new file, writing the directory structure to this file, and flushing it at the end.